Split models module up in smaller modules
